+++ tests/Zoop/Shard/Test/Serializer/TestAsset/Document/User.php
@@ -24,6 +24,7 @@ class User
     /**
      * @ODM\Field(type="string")
      * @Shard\Serializer\Ignore
+     * @Shard\Unserializer\Ignore
      */
     protected $password;
